new work on refactoring
