+++ OKKCNC/.spyproject/workspace.ini
@@ -6,5 +6,5 @@ save_non_project_files = False
 
 [main]
 version = '0.1.0'
-recent_files = [u'/home/carlo/Remote/OKKCNC/OKKCNC/Ribbon.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Commands.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/FilePage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/controllers/_GenericGRBL.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/OCV.py', u'/home/carlo/Remote/OKKCNC-leftovers/CNC.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/ControlPage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/CNCCanvas.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/CAMGen.py', u'/usr/lib/python2.7/lib-tk/Tkinter.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Utils.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Interface.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/CNC.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/TerminalPage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/__main__.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/EditorPage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Sender.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/ToolsPage.py']
+recent_files = [u'/home/carlo/Remote/OKKCNC/OKKCNC/CNCList.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/controllers/_GenericGRBL.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Commands.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/FilePage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/TerminalPage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/OCV.py', u'/home/carlo/Remote/OKKCNC-leftovers/CNC.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/ControlPage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/CNCCanvas.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/CAMGen.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/__main__.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Utils.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Sender.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Interface.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Ribbon.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/EditorPage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/GCode.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Block.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/Probe.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/ToolsPage.py', u'/home/carlo/Remote/OKKCNC/OKKCNC/ProbePage.py']
